requirements.txt
python-dotenv
streamlit
google-generativeai==0.3.2